extended pressure resolution to proper .1 of hPa

+++ system/include/aprs/wx.h
@@ -11,8 +11,8 @@
 #include "drivers/tx20.h"
 #include <stdint.h>
 
-void SendWXFrame(Anemometer* input, float temperatura, unsigned cisnienie);
-void SendWXFrameToBuffer(Anemometer* input, float temperatura, unsigned cisnienie, uint8_t* buffer, uint16_t buffer_ln, uint16_t* output_ln);
+void SendWXFrame(Anemometer* input, float temperatura, float cisnienie);
+void SendWXFrameToBuffer(Anemometer* input, float temperatura, float cisnienie, uint8_t* buffer, uint16_t buffer_ln, uint16_t* output_ln);
 
 
 #endif /* INCLUDE_APRS_WX_H_ */
